# specs/002-fastapi-todo-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API - Phase 2
  description: FastAPI backend for Todo application with JWT authentication from Next.js
  version: 1.0.0
  contact:
    name: Hackathon II Team
    url: https://github.com/anthropics/hackathon-todo
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todo-app.example.com
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /api/{user_id}/tags:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Better Auth user ID (must match JWT sub claim)
        example: "550e8400-e29b-41d4-a716-446655440000"

    get:
      tags:
        - tags
      summary: Get user's tags
      description: Returns all unique tags for the authenticated user with usage counts (for autocomplete)
      operationId: getUserTags
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tags with usage counts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagWithUsage'
              example:
                - name: "urgent"
                  usage_count: 5
                - name: "work"
                  usage_count: 3
                - name: "personal"
                  usage_count: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header (Bearer scheme)
  schemas:
...
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Task not found"
          description: Error message describing what went wrong
      required:
        - detail
...
    TagWithUsage:
      type: object
      required:
        - name
        - usage_count
      properties:
        name:
          type: string
          example: "urgent"
          description: Tag name
        usage_count:
          type: integer
          minimum: 0
          example: 5
          description: Number of tasks using this tag
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Missing or invalid authorization header"
    Forbidden:
      description: Not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authorized to access these tasks"
...